+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Support Hi-Res WAV up to 24bit / 192KHz
 * Support Mono WAV
 * Add Lcd Config in Config Menu
+* Force to reset ConfigParam when format revision changed to avoid mulfunction
 
 ### Changed
 * Use submodule for pico_audio_i2s_32b library